Extracting theme context to its own context and theme state management to its own hook
